(might be interesting for HEAD)
added statements to configure.in, so only shared libraries
are built in TNG. The opposite statement (building static
libraries by default) is commented out and interesting for
HEAD. This choosing can only be done in configure.in, not
at "configure"-time, as I hoped.

Also added the statement to let libtool check for dlopen.
This is the correct way of doing it, not modifying
ltconfig. (will roll back that soon)