configure configure.in: Fixed up errors in sambabook substitution that caused make install to fail. smbd/close.c: Interesting fix provoked by a customer problem when re-exporting nfs.
The nfs close() call was hanging for more than 30 seconds, and in the previous
version of this code the sysV semaphore was locked over this time. This was
causing other smbds to fail and get re-spawned. Moved the release of the
semaphore to just after the open mode is removed so the close() call is
no longer donw when the semaphore is held.

smbd/fileio.c: Paranioa fixes to ensure semaphore is released before abort()
called.

Jeremy.