"Wasted" this morning and last evening with researching:
Added lsa_enum_privs, client and server-side. This is a
list of stuff like SeMachineAccountPrivilege and some
sequential number, that starts with 2 and ends with 24.
Anyone any idea, what it is good for?
I was a little surprised, that one could do "enumprivs"
anonymously.

Also added lsa_priv_info, but only client-side, and the
rpcclient-command "privinfo" was just an attempt, I later
added better support directly in enumprivs (try
"enumprivs -i").

lsa_priv_info also has an (unaligned!) uint16 in its
response, that doesn't make any sense to me: For some
privilges it is 0x0, for others 0x407. (It's listed in
parenthesis after the description in enumprivs -i)