added a MANGLE_DRIVER_PATH define to chooose whether we stuff with the driver path do a rename etc.
I turned it off by default as it is causing me no end of problems. The
client will *cache* this path, and that screws things up badly as the
temp directory is no longer there when it is needed.
(This used to be commit 46f5f41c88b6ffa117d30f7702628b43dd456902)