I really didn't like the way generic arrays (lists) were implemented in util_array.c so I wrote a smaller (and simplier package).
I would like to replace the use of util_array.c functions
in the rest of the source tree if no one objects.
This will be an interface change, but not really a difference
in the functionality provided.


--jerry
(This used to be commit 0863ce3a31e4c76c8356850ac128f6d96ea54917)