Instead of handing back a string (which might be a DNS name or an IP string), the wins_srv module now hands back a struct in_addr when it's called.  It caches the IP address once it has been looked up.  The IP is cleared (and must be looked up again) if the 'wins server' parameter is reread, or if the node is marked 'dead'.  A dead node will not be re-tried for 10 minutes (per a #define in wins_srv.c).
As it was, the code was reading the WINS server name or IP directly from
lp_wins_server.  That's okay, except that if the value was expressed as
a name, then a DNS lookup would be done every time the client wanted to
talk to the server.

I still need to work out the implications of failover regarding the
'unicast subnet' list.

Chris -)-----
(This used to be commit 73aa188320fd3bf10b5dfc057323f40aff2c13bd)