First shot at actually *doing* WINS failover. If libsmb/namequery.c:name_query() times out while doing a non-broadcast query, I mark that WINS server 'dead'.  Note that I don't try the new WINS server.  I think I can get that working too.
This is only for queries, not registrations.  The biggest problem is that
I may have to fiddle with the UNICAST SUBNET, but I need to check talk
that over with someone (Jeremy?) before I hack at it.

I can't actually test the above change, I'm 'fraid.  I'm getting:

4 errors detected in the compilation of "rpc_server/srv_spoolss_nt.c".

in head branch.

Chris -)-----
(This used to be commit 40ae638866e3d826ccd5d827672634959c31b039)