fullname and description were exchanged on the wire in SAM_DISP_INFO_1.
Added info-level 6 for lsa_query_info_policy().
This one tells, if this machine is the primary or backup
controller for the domain it controls.
I added parsing and the server-side code.
This is needed for samba-as-pdc, w2k-in-workgroup,
user-list in acleditor.

w2k (at least the version, I had temporary access to) sends
some funky lsa_close-handle. Instead of typical 20 bytes,
the handle is only 2 bytes long and contains only a 01 00.

So I checked the return-value of every smb_io_pol_hnd in
parse_lsa.c, so lsarpcd could return a fault-pdu. (until
someone tells us, what's happening)