Fixing someone's problem where %U was not being substituted.
Not sure if this is the right fix ... Will take more code perusing
than I have done so far, and may need to look things up in a vuid?

Anyway, NET USE H: /HOME now works for Win9X clients ... Now for
Win2K which does not work. One of the RPCs is not working.