More fixes for floppy the rabbit.
Instead of being appliance only, the displaying of local users can now be
controlled by the smb.conf parameter 'hide local users'.  If true,
names and sids corresponding to local users on the samba server are not
displayed or resolved.  The default is false.

This may clash messily with the new passdb changes in HEAD. )-: