Add a new file clidgram with routines that are used by the libsmbclient code.
These routines handle the sending of dgrams in ways that don't bind us to
the nmbd code, but we may merge the two routines at some stage.

Also fix Makefile.in so the new code is compiled ...

Let's see whether or not it compiles on other architectures ...

Seems OK under Linux.