Fixed a bunch of printer permission bugs.
Implemented mapping of generic access rights to object specific rights for
NT printers as per the MSDN.  Basically any generic bits set in ACE masks
are mapped to object specific ones.  This unifies the observed ACE mask
values in security descriptors for NT4 and Windows 2000.  It also
simplified the code quite a bit.  Yay!

Unfortunately for the 'Manage Documents' and 'Full Control' permissions to
also allow 'Print' access to the printer, the check_ace() function needs to
*not* ignore ACEs with the SEC_ACE_FLAG_INHERIT_ONLY flag set.  Or maybe
not ignore ACEs with the SEC_ACE_FLAG_OBJECT_INHERIT also set.  I don't
know.  This should probably be nutted out.  See the comments in the
check_ace() function for more info.