Added a se_create_child_secdesc() function which takes a parent (container) security descriptor and creates a security descriptor which has the inheritance flags for each ACE applied.  In NT a print job is a child object of a printer so deleting and pausing/resuming jobs requires a check against the child security descriptor, not the parent.  The values seen in NT printer security descriptors now all fit together in a natural and elegant way which is always nice.
Removed #ifdef'ed out portion of check_ace() when the INHERIT_ONLY flag is
set as the se_create_child_secdesc() function now creates a security
descriptor which can be used without this hack.

NT trivia of the day: A user with Manage Documents privilege can't actually
print a document or open a printer.  Print permission is required to do
either of these tasks.