converted the passdb smbpasswd implementation to using talloc for memory allocation.  This fixes a long standing seg fault (i knew i would get around to it one day :) )
Tested with NT4 and Win2k.  Needs a little more testing with the
"create the machine account on the fly" code for NT4.

Simo, this is probably going to break the tdb passdb code.
Can you look at that when you get a chance and see what you think?
(This used to be commit 1c13110873e456748dc09fd51749f664643fe888)