Added back the automatic build of libsmbclient.so in head.
This is so I can find out what platforms it fails on. I will pull it again tomorrow if there are too many problems, like > 2 platforms that it fails to build on, but will pop it back in again as I resolve platforms.
(This used to be commit cc72c44e53a36f3da135f4dc10a9391f65c8b6c7)