Added info to some of the debug messages to get a better handle on a problem people are reporting regarding multiple responses to queries on <1D> names.
There should only ever be one LMB but some users are seeing multiple replies
to queries for the LMB name.  This is probably due to nodes on the LAN that
have NetBIOS over NetBEUI and/or IPX enabled.  Previously, the debug message
did not include the IP address associated with the name.  It *did* include
the source address of the packet, but in the examples I've seen all of these
were the same, eg:

[2000/06/22 11:58:25, 0] nmbd/nmbd_namequery.c:query_name_response(93)
query_name_response: Multiple (2) responses received for a query on subnet
129.130.10.136 for name NT.CIS.KSU.EDU<1d>. This response was from IP
129.130.10.24
 [2000/06/22 11:58:25, 0] nmbd/nmbd_namequery.c:query_name_response(93)
query_name_response: Multiple (3) responses received for a query on subnet
129.130.10.136 for name NT.CIS.KSU.EDU<1d>. This response was from IP
129.130.10.24
 [2000/06/22 11:58:25, 0] nmbd/nmbd_namequery.c:query_name_response(93)
query_name_response: Multiple (4) responses received for a query on subnet
129.130.10.136 for name NT.CIS.KSU.EDU<1d>. This response was from IP
129.130.10.24
[2000/06/22 11:58:25, 0] nmbd/nmbd_namequery.c:query_name_response(93)
query_name_response: Multiple (5) responses received for a query on subnet
129.130.10.136 for name NT.CIS.KSU.EDU<1d>. This response was from IP
129.130.10.24

Note that all of the above are reported as having come from 129.130.10.24.
This should never happen.  If 129.130.10.24 is a WINS server it should
send a Negative Name Query Response for a <1D> name query (wierd but true).

So, are all of the above coming from different systems, all of which
think are the LMB?  Are they all coming from one system that is, for some
strange reason, replying five times to the same query?

Anyway, I needed more info so I've changed the debug messages.

Chris -)-----
(This used to be commit 8f2f09af0a0a80cacef933ed500884e2c0b3f2fb)