oh my goodness!  He's gone an done it now! :-) Playing with LDAP support for 2.2 (non disruptively of course).
Initial ideas of storing smbpasswd in LDAP and then having
a cron job to export it every so often....

I even thought of possibly something like

	smb passwd file = |export_smbpasswd.pl

and having the smbpasswd file generate on the fly :-)

The point is that the full sam-db for accounts is a long ways off
in theory.  This will make some people happy.  The OIDs used in
the schema are ours.


jerry