This fix from Eelco Vriezekolk <eelco@nexus.com.na> is for a SIG11 bug where we would always file_free(fsp) twice, once in close_file and once afterwoulds.
The bug was reported in SAMBA_2_2, but a code inspection shows it to be in HEAD
as well.  (Unfortunetly I don't have the facilites to actualy check this, but
the change is quite simple, makes sence and compiles).

Andrew Bartlett
(This used to be commit 5d3f622375842961a79536d36ee6f6ad508b91f1)