this fixes the failure of MS office on VFAT partitions on Linux
The problem is that ftruncate can't expand on VFAT, but it can on
ext2. That means our autoconf test is useless. I have recoded it to
use the alternative to ftruncate when then sys_ftruncate fails.

Jeremy, do you want this for 2.2.1?
(This used to be commit 970236ee9926b64c0b39bd8a36b2a9317206873c)