Fix the loading of configuration files using the include syntax.
We had a problem where if a % macro in the smb.conf could be ignored if the
various files it pointed to had the same time-stamp.

This changes the code to insted check that the both the time-stamp and the
substituted filename are the same over each change.

This was picked up only becouse the build-farm automaticly generates its config
files, and hence gets identical timestamps.

(Why this doesn't happen all the time I'm not entirly sure, somthing to do with
the 'test' paramater to reload_services(), but this fixes this problem).

Andrew Bartlett
(This used to be commit ebd2f9b07c89cce505e821f1caaa6817bbb26db9)