Store winbindd in the sbin directory.  Make the winbind pam module also as a sbin program.
Currently the pam and nss modules are installed into @prefix@/sbin - I'm
not sure whether this is a good idea or not.  Perhaps they should be left
in the build tree and copied across as needed by hand or a packaging tool.
(This used to be commit 09af934c4aa295bec494740d94de76e5e4497887)