ok.  This introduces the 'lanman printing only' parameter which disables the spoolss pipe.  This provides us with full backwards compatibility with Samba 2.0.x and NT printing clients.
The reason for this patch is that an NT print server cannot
have a printer with no driver installed.  Samba can, so NT clients
get confused because they have never seen that before.  So now
we start installing local drivers for printers served using MS-RPC.
But the NT client tries to issue an open_printer_ex() call with
the printer rights of the locally logged in user (e.g.
the local administrator would try to open this printer
with PRINTER_ADMINISTER_ACCESS).  If this fails, the client
puts a nice "access denied. unable to connect" message in the title
bar of the print queue window.

Jeremy, i know we need to talk about this.  See api_DosPrintQGetInfo()
in smbd/lanman.c.