lp_wins_server() is now lp_wins_server_list().
This should make it clear that the meaning of the parameter value has
changed.  It no longer represents *the* WINS server, but a list of WINS
servers.

I have made other changes in the code such that the lp_wins_server()
function is no longer necessary.  Whenever smb.conf is reloaded the list
managed by lib/wins_srv.c is refreshed.  The wins_srv_count() function
returns the number of entries in the list so, if the list is empty, it
will return 0 (which can be interpreted as "false" in an if() statement).

Chris -)-----