Fuss, fuss, fuss...
Function name_status_query() performs a node status query, so it really
should be called "node_status_query()" just to be consistent.

Yeah, minor... but it's looking as though an overhaul of namequery.c is in
order and I am trying to do the tiny changes that impact other stuff first.

Chris -)-----