Reverse some of the breakage I commited a day or two ago, as we need to fill out the user_info struct (otherwise we don't have a vuid for make_connection()).
Also add a become_user() call, becouse it really looks like it was missing
(we must pass it anyway to finish make_connection()).

Is there any reason not to be the user when reading an ACL?

Finally, fix up some formatting to show that the two functions are almost
identical.

Andrew Bartlett
(This used to be commit 00c667c0ad922a1bf388b8a2b8c6137fc7f0acaa)