Fix for universal groups printing bug: if a user is a member of group that is a member of a universal group, and that universal group has a deny ACE on a printer security descriptor Samba fails to deny the user access.
I've passed up the group rids from the cli_login_network() response and
added to the nt user token when it is created.

This works for printing but the groups are lost when connecting to
a share due to the token be recreated due to the possible use of
the 'force user' and 'force group' parameters.