This change updates lp_guestaccount() to be a *global* paramater, rather than per-share.  I beleive that almost all the things that this could have done on a per-share basis can be done with other tools, like 'force user'.
Almost all the user's of this paramater used it as a global anyway...

While this is one step at a time, I hope it will allow me to considerably
simplfy the make_connection() code, particularly for the user-level security
case.

This already removes an absolute truckload of extra attempted password lookups
on the guest account.

Andrew Bartlett